pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "archml"
version = "0.1.0"
description = "A text-based DSL for defining software architecture alongside code"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = [
    "dash>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
archml = "archml.cli.main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.9",
    "ty>=0.0.1a6",
    "yachalk>=0.1.8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.hatch.build.targets.wheel]
packages = ["src/archml"]
